Move use of transcription to dfs cggnn call.
